refactor(footer): restructure Footer component into modular subcomponents
Split monolithic Footer component into organized directory structure:
- Create footer/ directory with Footer, StateControls, and Pages components
- Extract shared FooterFrame wrapper component for consistent styling
- Separate state management controls (theme, animations, chat) into StateControls
- Add Pages component for future page navigation functionality
- Update import path in page.tsx to reference new footer structure

This improves component organization, maintainability, and enables
future expansion of footer functionality without bloating a single file.
